[refactor] - entity long -> Long type 으로 변경 - TokenStatusResult dto 추가 - queue infrastructure 패키지 추가 - 만료 스케줄러 주기 1분으로 변경

+++ src/main/java/kr/hhplus/be/server/api/reservation/application/dto/ReservationDto.java
@@ -11,7 +11,7 @@ public record ReservationDto(
 	long id,
 	long userId,
 	long seatId,
-	long seatNo,
+	int seatNo,
 	LocalDate concertDate,
 	BigDecimal finalPrice
 ) {
